Fork the favorites extension.
That's fork in terms of "copy" and not in terms of "git clone and set
upstream".  The extension comes from

https://www.mediawiki.org/wiki/Extension:Favorites

and was pulled from master at

https://gerrit.wikimedia.org/r/plugins/gitiles/mediawiki/extensions/Favorites/+/master

The commit was c4546327c7669fd6cc7344e34c2c96a1d1d86aa0

This fork is to lay the groundwork to cull a lot of things that don't
work for torque out.