Fork the Comments extension.
That's fork in terms of "copy" and not in terms of "git clone and set
upstream".  The extension comes from

https://www.mediawiki.org/wiki/Extension:Comments

and was pulled from master at

https://gerrit.wikimedia.org/g/mediawiki/extensions/Comments

The commit was 	a99e6eec99e6b50a90856cb26b9a864fad31a949

This fork is to lay the groundwork to cull a lot of things that don't
work for torque out, and update for extra features we want.