Add configuration to allow locallogin override.
The target wiki, when using okta, needs to both allow local logins and
force users to okta.  To achieve this, we hide a flag that overrides the
redirect to okta if you know what you're doing.  This doesn't create a
security risk because you still need to be able to log in via
credentials on the target wiki.