Split out csv2wiki and mwlib users.
There are two external apps that use the mediawiki instance, the
population, and the mwlib app for generating pdfs.  They were using
one account for both, and that needed to be changed to separate.

This also makes the mwlib user not a bot so that it doesn't have write
access.