Add mwiki template rendering.
This is an intermediary commit that has some hard coded values, such as
the template to render, as well as the permissions for the json api.
Putting it in place to develop the mediawiki plugin side of this feature
before moving on to more dynamic configuration.

The mwiki.j2 file was generated based on the config file used in
macfound.