Incremental move to data agnostic torque data.
Instead of hardcoded proposals as a thing, this change allows uploading
of arbitrary spreadsheets along with configuration of them, which when
combined with future changes will make torquedata completely agnostic of
what's in it.  The changes mainly have to do with how routes are
configured.

There's still some holdouts to the previous hardcoded version, but those
will go away soon.