Add group lookup for columns, readd attachments.
This accidentally got added in one commit, and was too hard to untangle.
Half of the commit is adding groups, and looking up the groups, then
culling the columns by the group looked up.  The other half is
adding attachments back in in a hardcoded method.

A note on the hardcoding.  Right now we're adding all the needed mwiki
edit features back in as hardcoded items in order have a sense of what
general solution is going to work best.