Set stage for toc uploads.
Not only is the routes table added for it, but some refactoring needed
to happen to clear the way.  Mainly the spreadsheets directory needs to
have a deeper directory heirarchy.  While in there, also changed plural
to sheet_name and singular to object_name to be clearer from the point
of view of the code rather than the outside usage.