Fix Book Printing with TorqueData.
Because torquedata renders based on what group the user is, that group
has to be propogated through to the mwlib server.  This required
patching mwlib, and then managing the hooks in mediawiki so that the
group is set in the right places.

This also included a change to pull groups from the config instead of
having them be hardcoded in the TorqueDataConnect extension.