Fix book printing.
In order for book printing to work correctly, we need a top level switch
between rendering a page as html, or rendering it as wiki text.  In the
case of html, we can further render any internal hooks, but that breaks
how mwlib was rendering pages.

There's also a small fix that got snuck in about setting the error
message when a key was not found.