Add raw rendering, wiki_key.
The goal was adding a second wiki that was a skinned, public
site of the same data as a main wiki.  In order to do this, we needed
the new wiki to have the ability to do two things.

The first is disabling the wiki rendering in the TorqueDataConnect
pipeline, via the new variable $wgTorqueDataConnectRaw

The second is that we needed the two wikis to have unique config in the
torque engine, which required them to each have an identifier that they
pass for each request.