Add basic concept of users to torque, create lookup
We need a centralized list of users so that all competitions can refer
to the same set of people, indexed by username.  This adds a starting
point for that, as well as a extension class that MediaWiki can use to
access it.

This should probably be abstracted in the future to some kind of user
class that exists in a database.