Reload the current page when editing
Before, we were returning just the default template view for the
proposal we happened to be editing.  This broke cases where the view
that the wiki was using was something with extra wiki text, or pages
where there were multiple proposals, or when we were using an alternate
view for the proposal.

This also required changing the tdcrender code to pull the group at the
last second from the user if none was set.  The parser, when rendering
the page, wasn't running all the hooks (as expected), so the system
wasn't set up as needed.