Switch torquedata entirely to Django
As per discussion in issue #26 and over email correspondence, the plan
was to keep torquedata in Flask but use the Django ORM to interface with
postgres, keeping the option of migrating entirely to Django in the
future.

In practice, this presented some nontrivial issues. Since the data model
effectively *is* the application, torquedata would effectively become a
Django project that happens to use Flask to handle views and routing.
Support for non-standard Django projects online is minimal and
ironically requires more specific knowledge of Django's inner workings.
Torquedata's views are so dependent on the old pickling data model that
they will all have to be rewritten anyways, which means sticking with
Flask is an unnecessary complication.

A new Django app was created with `django-admin startproject`, and the
models and management commands were copied over.