When loading a new sheet, append new edit entries
There was an issue where, if we loaded a sheet with new columns, there
would be no edits record for those new columns, causing the renderer to
break when it attempted to apply edits to that new column.  This changes
adds those entries to the edits object if they don't already exist.

There was another potential issue where a new sheet would not get added
to the edits database if the database had been loaded from disk and the
new sheet was uploaded.  This should also fix that.