Change search to use the current wiki's template
Originally search was going to use the template to render search results
that was based on the wiki of the result itself.  However, that toc will
always reference pages within the internal wiki, and may have a
different look and feel to other wikis.

In order to link tot he correct pages in the global view wiki, and in
order to make sure all search results look alike, there must be a
template set up for the combination wiki.  This also required changing
the api to take an additional parameter for the wiki making the call.