Small fixes to TOC upload code
Change it to be create_or_update for re-uploaded TOCs, as well as
enabling and using the None type for the templates associated with the
TOC at upload time.

Background information for above, is that the "TOC" type for templates
is reserved for the template that is set up by the configuration of the
wiki in the MainConfig, which is responsible for creating lines in the
table of contents.