Add flags for config cache and sha
This exists so that we don't rebuild the search cache when we aren't
actually updating the columns/rows that groups have permissions to.
This also changes which rows get processed as SearchVectors, limited to
those that were pertinent.

This is done through a cache of a sha created based on the proposal and
column names.