Update group lookup
We were using two deprecated methods to look up the user group.  The
first was the use of the global `$wgUser` which is considered unsafe, as
it may not be instantiated at any given moment.

The second was the use of `->getGroups()` which was replaced in 1.35
with a UserGroupManager class.

Issue #56