Cache Tables of Contents in HTML
This required calling out to the target wiki with the wiki text, and
storing that instead of the rendered mwiki format.  Then, when the
caller calls, they can call with html if they want the cached version,
or mwiki with the fresh version.

Right now it hardcodes the wiki it contacts, and the server for
development purposes, so this commit doesn't fully work.  That needs to
be parameterized (the next commit).