Use json instead of csv for uploaded sheet
It turns out that we don't need to actually upload a sheet, and then
deal with that, and it's better to just upload a json document that we
can turn into a dict and work with.  We then store each field in the
database in json form, before pulling back out and turning into python
objects before passing to the templates.