Add ignorable filters, pass dictionary to filter
The filter processing code had a bug where the full Document was getting
sent, irrespective of permissions.  This fixes that.

It also adds the concept of values that should get ignored, which can
lead to filters getting ignored.  This is mainly used for permission
structures where someone shouldn't see a filter that they don't have
permission to really use.