Don't delete templates from database on config reset
This required a in_config flag, as well as passing whether it's default
from mediawiki.  The reason for this change is now that we've added
caching for templates, we need to not blow away the cache unless the
template is removed from the configuration.  A simple save on the
mediawiki config page was deleting the whole cache.